+++ mdObject/src/test/classes/Allergy.spec.ts
@@ -1,4 +1,4 @@
-import { Allergy } from "../../classes/classes";
+import { Allergy } from "../../classes";
 import { MockEmrMel } from "../mocks/mocks";
 import { AllergyClassification, AllergyCriticality, ObjectState } from "../../enums/enums";