+++ mdObject/src/test/classes/PatientContactAddress.spec.ts
@@ -1,4 +1,4 @@
-import { PatientContact } from '../../classes/classes';
+import { PatientContact } from '../../classes';
 import { mockResultAddress } from '../mocks/mocks';
 
 describe('Class: PatientContactAddress', () => {