+++ mdObject/src/test/classes/User.spec.ts
@@ -1,4 +1,4 @@
-import { User } from '../../classes/classes';
+import { User } from '../../classes';
 import { mockResultEmr } from '../mocks/mocks';
 import { UserCallFunction } from '../../enums/enums';