+++ mdObject/src/classes/Problems.ts
@@ -2,7 +2,7 @@ import { Problem } from "./Problem";
 import { StringInternal } from '../factories/factories';
 import { EmrMel } from "./EmrMel";
 import { ArrayAsync } from "../interfaces/ArrayAsync";
-import { ObjectStatus } from "../enums/enums";
+import { ObjectStatus } from "../enums";
 
 export class Problems extends ArrayAsync<Problem>{
     private _isLoaded = false;