+++ mdObject/src/classes/Users.ts
@@ -1,7 +1,7 @@
 import { User } from './User';
 import { EmrMel } from './EmrMel';
 import { StringInternal, GetCurrentUser } from '../factories/factories';
-import { UserCallFunction } from '../enums/enums';
+import { UserCallFunction } from '../enums';
 import { IArrayAdditionalMethods } from '../interfaces';
 
 export class Users {