+++ mdObject/src/test/classes/Location.spec.ts
@@ -1,5 +1,5 @@
 import { Location } from '../../classes';
-import { LocationType } from '../../enums/enums';
+import { LocationType } from '../../enums';
 
 describe('Class: Location', () => {